% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttackFunction.R
\name{attack}
\alias{attack}
\title{Roll an Attack in Dungeons and Dragons}
\usage{
attack(mod, ac)
}
\arguments{
\item{mod}{(this is your attack modifier)}

\item{ac}{(this is the enemy AC)}
}
\value{
Prints "hit!" or "miss!" depending on if the total "attack roll" is >= the AC of the "monster"
}
\description{
Roll an Attack in Dungeons and Dragons
}
\examples{
attack(3, 15)

}
